campus-cv/requirements.txt
anthropic==0.19.2
Flask==3.0.2
fpdf==1.7.2
psycopg2-binary